Add files via upload
I am working on this script again to add a limited reactive transport capability (a couple weeks out, still). In the meantime, while reviewing it, I found and fixed a couple of bugs. One had to do with the automatic grid generator when cells are duplicated - I hadn't tested this feature before and it didn't work. Fixed it. The second entailed a subtle error involving attaching an array to a data structure that gave unintended results for gas diffusion problems. Fixed that as well. More later ...